Update fix-error-for-submission-quota-exceeded-in-exchange-online.md
Removed this information because it is not accurate for users that reach their recipient rate limit - 
Go to the Restricted users portal in the Security & Compliance Center at https://protection.office.com/?hash=restrictedusers to unblock the account. After you unblock the account, the user should be able to resume sending messages within a few hours.