Update fix-error-for-submission-quota-exceeded-in-exchange-online.md
Removed this section which is not valid for this type of failure. Senders will have to wait 24 hours to send mail if they hit their sending limit - 

- Go to the **Restricted users** portal in the Security & Compliance Center at <https://protection.office.com/?hash=restrictedusers> to unblock the account. After you unblock the account, the user should be able to resume sending messages *within a few hours*.

Also Microsoft is introducing a more friendly NDR message that helps users understand why the message was blocked and that they will need to wait 24 hours to send more email. This includes the NDR code 5.1.92.